+++ lib/acvp_list_requests.c
@@ -23,12 +23,9 @@
 #include "definition.h"
 #include "internal.h"
 
-static void acvp_list_request_id(const uint32_t id, const char *module_name,
-				 const char *name, const uint32_t testid)
+static void acvp_list_id(const uint32_t id, const char *module_name,
+			 const char *name, const uint32_t testid)
 {
-	if (acvp_valid_id(id))
-		return;
-
 	if (id == 0) {
 		if (logger_get_verbosity(LOGGER_C_ANY) >= LOGGER_DEBUG)
 			fprintf(stdout, "%-40s | %-8u | %-15s | no ID\n",
@@ -39,20 +36,31 @@ static void acvp_list_request_id(const uint32_t id, const char *module_name,
 	}
 }
 
-static void acvp_list_id(const uint32_t id, const char *module_name,
-			 const char *name, const uint32_t testid)
+static void acvp_list_request_id(const uint32_t id, const char *module_name,
+				 const char *name, const uint32_t testid)
+{
+	if (acvp_valid_id(id))
+		return;
+
+	acvp_list_id(id, module_name, name, testid);
+}
+
+static void acvp_list_avail_id(const uint32_t id, const char *module_name,
+			       const char *name, const uint32_t testid)
 {
 	if (!acvp_valid_id(id))
 		return;
 
-	fprintf(stdout, "ID for %s (test session %u) - %s: %u\n", module_name,
-		testid, name, id);
+	acvp_list_id(id, module_name, name, testid);
 }
 
 static int acvp_list_certificate_id(const struct acvp_ctx *ctx,
 				    const struct definition *def,
 				    const uint32_t testid,
-				    const bool list_request)
+				    void(*list_func)(const uint32_t id,
+						     const char *module_name,
+						     const char *name,
+						     const uint32_t testid))
 {
 	const struct def_info *def_info;
 	struct acvp_testid_ctx testid_ctx;
@@ -73,36 +81,21 @@ static int acvp_list_certificate_id(const struct acvp_ctx *ctx,
 	/* Get testsession ID */
 	auth = testid_ctx.server_auth;
 
-	if (list_request) {
-		acvp_list_request_id(auth->testsession_certificate_id,
-				     def_info->module_name, "certificate",
-				     testid);
-	} else {
-		acvp_list_id(auth->testsession_certificate_id,
-			     def_info->module_name, "certificate", testid);
-	}
+	list_func(auth->testsession_certificate_id, def_info->module_name,
+		  "certificate", testid);
 
 out:
 	acvp_release_auth(&testid_ctx);
 	return ret;
 }
 
-static int acvp_list_certificate_id_cb(const struct acvp_ctx *ctx,
-				       const struct definition *def,
-				       const uint32_t testid)
-{
-	return acvp_list_certificate_id(ctx, def, testid, false);
-}
-
-DSO_PUBLIC
-int acvp_list_certificate_ids(const struct acvp_ctx *ctx)
-{
-	return acvp_process_testids(ctx, &acvp_list_certificate_id_cb);
-}
-
-static int acvp_list_requests_cb(const struct acvp_ctx *ctx,
-				 const struct definition *def,
-				 const uint32_t testid)
+static int acvp_list_ids_cb(const struct acvp_ctx *ctx,
+			    const struct definition *def,
+			    const uint32_t testid,
+			    void(*list_func)(const uint32_t id,
+					     const char *module_name,
+					     const char *name,
+					     const uint32_t testid))
 {
 	struct def_info *def_info;
 	struct def_vendor *def_vendor;
@@ -115,37 +108,58 @@ static int acvp_list_requests_cb(const struct acvp_ctx *ctx,
 	def_vendor = def->vendor;
 	def_oe = def->oe;
 
-	CKINT(acvp_list_certificate_id(ctx, def, testid, true));
+	CKINT(acvp_list_certificate_id(ctx, def, testid, list_func));
 
 	CKINT(acvp_def_get_vendor_id(def_vendor));
-	acvp_list_request_id(def_vendor->acvp_vendor_id, def_info->module_name,
-			     "vendor", testid);
-	acvp_list_request_id(def_vendor->acvp_addr_id, def_info->module_name,
-			     "address", testid);
+	list_func(def_vendor->acvp_vendor_id, def_info->module_name, "vendor",
+		  testid);
+	list_func(def_vendor->acvp_addr_id, def_info->module_name, "address",
+		  testid);
 	CKINT(acvp_def_put_vendor_id(def_vendor));
 
 	CKINT(acvp_def_get_person_id(def_vendor));
-	acvp_list_request_id(def_vendor->acvp_person_id, def_info->module_name,
-			     "person", testid);
+	list_func(def_vendor->acvp_person_id, def_info->module_name, "person",
+		  testid);
 	CKINT(acvp_def_put_person_id(def_vendor));
 
 	CKINT(acvp_def_get_module_id(def_info));
-	acvp_list_request_id(def_info->acvp_module_id, def_info->module_name,
-			     "module", testid);
+	list_func(def_info->acvp_module_id, def_info->module_name, "module",
+		  testid);
 	CKINT(acvp_def_put_module_id(def_info));
 
 	CKINT(acvp_def_get_oe_id(def_oe));
-	acvp_list_request_id(def_oe->acvp_oe_id, def_info->module_name,
-			     "OE", testid);
-	acvp_list_request_id(def_oe->acvp_oe_dep_sw_id, def_info->module_name,
-			     "SW dependency", testid);
-	acvp_list_request_id(def_oe->acvp_oe_dep_proc_id,
-			     def_info->module_name, "Proc dependency",
-			     testid);
+	list_func(def_oe->acvp_oe_id, def_info->module_name, "OE", testid);
+	list_func(def_oe->acvp_oe_dep_sw_id, def_info->module_name,
+		  "SW dependency", testid);
+	list_func(def_oe->acvp_oe_dep_proc_id, def_info->module_name,
+		  "Proc dependency", testid);
 	CKINT(acvp_def_put_oe_id(def_oe));
 
 out:
 	return ret;
+}
+
+static int acvp_list_avaiable_id_cb(const struct acvp_ctx *ctx,
+				    const struct definition *def,
+				    const uint32_t testid)
+{
+	return acvp_list_ids_cb(ctx, def, testid, &acvp_list_avail_id);
+}
+
+DSO_PUBLIC
+int acvp_list_available_ids(const struct acvp_ctx *ctx)
+{
+	fprintf(stdout, "%-40s | %-8s | %-15s | %-8s\n",
+			"Module Name", "Test ID", "ID Type", "ID");
+
+	return acvp_process_testids(ctx, &acvp_list_avaiable_id_cb);
+}
+
+static int acvp_list_requests_cb(const struct acvp_ctx *ctx,
+				 const struct definition *def,
+				 const uint32_t testid)
+{
+	return acvp_list_ids_cb(ctx, def, testid, &acvp_list_request_id);
 }
 
 DSO_PUBLIC